fix(arc-1408): move auth modal to single location: app layout
https://meemoo.atlassian.net/browse/ARC-1408
