+++ src/modules/shared/layouts/AppLayout/AppLayout.tsx
@@ -4,10 +4,14 @@ import { useRouter } from 'next/router';
 import { FC, useCallback, useEffect, useMemo } from 'react';
 import { useSelector } from 'react-redux';
 import { Slide, ToastContainer } from 'react-toastify';
+import { BooleanParam } from 'serialize-query-params/lib/params';
+import { StringParam, useQueryParams } from 'use-query-params';
 
 import { Permission } from '@account/const';
+import { AuthModal } from '@auth/components';
 import { AuthService } from '@auth/services/auth-service';
 import { checkLoginAction, selectIsLoggedIn, selectUser } from '@auth/store/user';
+import { SHOW_AUTH_QUERY_KEY, VISITOR_SPACE_SLUG_QUERY_KEY } from '@home/const';
 import { Footer, Navigation, NavigationItem } from '@navigation/components';
 import {
 	footerLeftItem,
@@ -29,6 +33,7 @@ import ErrorBoundary from '@shared/components/ErrorBoundary/ErrorBoundary';
 import { useGetNotifications } from '@shared/components/NotificationCenter/hooks/get-notifications';
 import { useMarkAllNotificationsAsRead } from '@shared/components/NotificationCenter/hooks/mark-all-notifications-as-read';
 import { useMarkOneNotificationsAsRead } from '@shared/components/NotificationCenter/hooks/mark-one-notifications-as-read';
+import { SEARCH_QUERY_KEY } from '@shared/const';
 import { WindowSizeContext } from '@shared/context/WindowSizeContext';
 import { useHasAllPermission } from '@shared/hooks/has-permission';
 import { useHistory } from '@shared/hooks/use-history';
@@ -40,6 +45,7 @@ import { getTosAction } from '@shared/store/tos/tos.slice';
 import {
 	selectHasUnreadNotifications,
 	selectIsStickyLayout,
+	selectShowAuthModal,
 	selectShowFooter,
 	selectShowNavigationBorder,
 	selectShowNotificationsCenter,
@@ -61,6 +67,7 @@ const AppLayout: FC = ({ children }) => {
 	const user = useSelector(selectUser);
 	const sticky = useSelector(selectIsStickyLayout);
 	const showFooter = useSelector(selectShowFooter);
+	const showAuthModal = useSelector(selectShowAuthModal);
 	const showNotificationsCenter = useSelector(selectShowNotificationsCenter);
 	const hasUnreadNotifications = useSelector(selectHasUnreadNotifications);
 	const windowSize = useWindowSize();
@@ -72,6 +79,11 @@ const AppLayout: FC = ({ children }) => {
 	const canManageAccount = useHasAllPermission(Permission.MANAGE_ACCOUNT);
 	const showLinkedSpaceAsHomepage = useHasAllPermission(Permission.SHOW_LINKED_SPACE_AS_HOMEPAGE);
 	const linkedSpaceSlug: string | null = user?.visitorSpaceSlug || null;
+	const [query, setQuery] = useQueryParams({
+		[VISITOR_SPACE_SLUG_QUERY_KEY]: StringParam,
+		[SEARCH_QUERY_KEY]: StringParam,
+		[SHOW_AUTH_QUERY_KEY]: BooleanParam,
+	});
 
 	useHistory(asPath, history);
 
@@ -109,11 +121,34 @@ const AppLayout: FC = ({ children }) => {
 		dispatch(getTosAction());
 	}, [dispatch]);
 
+	// Sync showAuth query param with store value
+	useEffect(() => {
+		if (user) {
+			setQuery({
+				...query,
+				[SHOW_AUTH_QUERY_KEY]: undefined,
+			});
+			dispatch(setShowAuthModal(false));
+		} else if (typeof query.showAuth === 'boolean') {
+			dispatch(setShowAuthModal(query.showAuth));
+		}
+	}, [dispatch, query.showAuth, user]);
+
 	const userName = (user?.firstName as string) ?? '';
 
 	const onLoginRegisterClick = useCallback(() => dispatch(setShowAuthModal(true)), [dispatch]);
 
 	const onLogOutClick = useCallback(() => AuthService.logout(), []);
+
+	const onCloseAuthModal = () => {
+		if (typeof query[SHOW_AUTH_QUERY_KEY] === 'boolean') {
+			setQuery({
+				[SHOW_AUTH_QUERY_KEY]: undefined,
+				[VISITOR_SPACE_SLUG_QUERY_KEY]: undefined,
+			});
+		}
+		dispatch(setShowAuthModal(false));
+	};
 
 	const rightNavItems: NavigationItem[] = useMemo(() => {
 		if (isLoggedIn) {
@@ -138,8 +173,8 @@ const AppLayout: FC = ({ children }) => {
 
 		return NAV_ITEMS_RIGHT(onLoginRegisterClick);
 	}, [
-		isLoggedIn,
 		onLoginRegisterClick,
+		isLoggedIn,
 		canManageAccount,
 		asPath,
 		navigationItems,
@@ -256,6 +291,8 @@ const AppLayout: FC = ({ children }) => {
 			/>
 
 			<ZendeskWrapper />
+
+			<AuthModal isOpen={showAuthModal && !user} onClose={onCloseAuthModal} />
 
 			{showFooter && (
 				<Footer
